refactor : ImageUtil 클래스를 Global.Image 패키지 하위로 이동
