+++ src/main/java/app/bottlenote/review/service/ReviewService.java
@@ -3,7 +3,7 @@ package app.bottlenote.review.service;
 import app.bottlenote.alcohols.domain.Alcohol;
 import app.bottlenote.alcohols.domain.AlcoholQueryRepository;
 import app.bottlenote.alcohols.exception.AlcoholException;
-import app.bottlenote.global.ImageUtil;
+import app.bottlenote.global.image.ImageUtil;
 import app.bottlenote.global.service.cursor.PageResponse;
 import app.bottlenote.review.domain.Review;
 import app.bottlenote.review.domain.ReviewImage;
